Remove FarmHash from the comparisons
FarmHash requires the entire output to be provided in one shot, making
it fundamentally incompatible with Rust's hashing design. To work around
this, the hasher implementation buffers the entire input until the hash
is finalized.

Including it here serves no purpose as it's the wrong algorithm for the
problem, which incorrectly makes an excellent algorithm look poor.